+++ src/ResourceManager/Resources/Commands.Resources/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Add new cmdlet Get-AzureRmPolicyAliases for retrieving policy aliases
 
 ## Version 6.4.2
 * Fixed issue with creating managed applications from the MarketPlace.